Keep both test key fixtures in one conftest
The SSH key fixture sat in the server conftest to avoid importing paramiko for
non-server runs, but the root conftest already pulls it in transitively, so there
was nothing to avoid. Placing both together also covers the CLI fleet
configurator and drops the order-dependent patching.
